Move symptom checker logic to display conditions file.
